Added "video support"
In the console interface, everything is now natively clear. The script itself understands what is in front of it (photo or video). Also, the ability to process all photos in a directory at once has been added. Semaphores have been added to perform several tasks asynchronously. The number of semaphores is also adjustable. For video, the user also sets the number of fps.